Added poll and member commands
📜 New functions 📜
✨ 'Poll': creates poll with given title, lasting given time 🗳
Parameters: 
duration - duration of the poll, followed by the time unit (s/m/h or d)
title - title of the poll
✨ 'Member' => 'Roles': Remake of old command. Added parameter, which stands for role name 🎭
Grants or revokes role, according to reaction to the bot message 🙋‍♂️ 
For now accessible only for 3 people on our server. Might be useful to give user access to intresting channels and granting appropriate roles (with permission to view those channels).