Added role command
(Correction of the recent push, 'Role' command was created in this version)
✨ 'Member' => 'Roles': Remake of old command. Added parameter, which stands for role name 🎭
Grants or revokes role, according to reaction to the bot message 🙋‍♂️ 
For now accessible only for 3 people on our server. Might be useful to give user access to intresting channels and granting appropriate roles (with permission to view those channels).