Remake of newMember command - roles are no longer needed
Remake of 'newMember' command so that instead of restricting access to text channels with a role, it does so directly for the user. The server will not be cluttered with unnecessary roles for each new member. The onUserJoin function has also been added, which should run the 'newMember' command and create a workflow as soon as the user joins the server 🏁